refactor: 증상 선택 페이지 로직 커스텀훅 분리
