+++ src/pages/symptom-type/lib/styles.tsx
@@ -1,7 +1,3 @@
-import styled from "styled-components";
-
-const Wrapper = styled.div``;
-
 const LAYOUT_STYLE: React.CSSProperties = {
   minHeight: "100%",
   background: "var(--gradient)",
@@ -9,6 +5,5 @@ const LAYOUT_STYLE: React.CSSProperties = {
 };
 
 export default {
-  Wrapper,
   LAYOUT_STYLE,
 };
